+++ src/core/Account/Handlers/Clear.cs
@@ -1,6 +1,5 @@
 using System.Threading;
 using System.Threading.Tasks;
-using core.Adapters.Emails;
 using core.Shared;
 using MediatR;
