+++ src/core/Account/Handlers/Delete.cs
