+++ src/core/Account/Handlers/UserConfirmedHandler.cs
