+++ src/core/Account/Handlers/UserCreatedHandler.cs
