--- src/core/Account/UserConfirmed.cs
@@ -1,13 +1,0 @@
-using System;
-using core.Shared;
-using MediatR;
-
-namespace core.Account
-{
-    public class UserConfirmed : AggregateEvent, INotification
-    {
-        public UserConfirmed(Guid id, Guid aggregateId, DateTimeOffset when) : base(id, aggregateId, when)
-        {
-        }
-    }
-}
