--- src/core/Account/UserLoggedIn.cs
@@ -1,12 +1,0 @@
-using System;
-using core.Shared;
-
-namespace core.Account
-{
-    internal class UserLoggedIn : AggregateEvent
-    {
-        public UserLoggedIn(Guid id, Guid aggregateId, DateTimeOffset when) : base(id, aggregateId, when)
-        {
-        }
-    }
-}
